% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{rexcel_register}
\alias{rexcel_register}
\title{Register an Excel workbook}
\usage{
rexcel_register(path)
}
\arguments{
\item{path}{path to xlsx}
}
\value{
a list
}
\description{
Experimental function, in an experimental package! It covers alot of the same
ground as \code{\link{rexcel_read_workbook}()} with several notable
exceptions.
}
\details{
\itemize{
\item Currently returns a list, not a proper linen::workbook.
\item Read one file at a time, make one or more objects from it, with only
the processing necessary to make it a reasonable R object.
\item For some clearly useful things, create downstream objects via
processing and/or combining info across primary objects.
}
}
\examples{
mini_gap_path <- system.file("sheets", "mini-gap.xlsx", package = "rexcel")
rexcel_register(mini_gap_path)

ff_path <- system.file("sheets", "gs-test-formula-formatting.xlsx",
                       package = "rexcel")
rexcel_register(ff_path)
}
\keyword{internal}

