% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{rexcel_workbook}
\alias{rexcel_workbook}
\title{Low-level function to expose contents of xlsx}
\usage{
rexcel_workbook(path)
}
\arguments{
\item{path}{}
}
\value{
a list
}
\description{
This is just Jenny getting to know xlsx! Function returns alot of the same
information as rexcel_read_workbook() but with several notable exceptions.
Returns a list, not a proper linen::workbook. Much less processing is done --
basically only whats needed to some reasonable R object, usually a data
frame.
}
\examples{
mini_gap_path <- system.file("sheets", "mini-gap.xlsx", package = "rexcel")
rexcel_workbook(mini_gap_path)

ff_path <- system.file("sheets", "gs-test-formula-formatting.xlsx",
                       package = "rexcel")
rexcel_workbook(ff_path)
}
\keyword{internaln}

