% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{xlsx_read_file}
\alias{xlsx_read_file}
\title{Read XML for a specific file}
\usage{
xlsx_read_file(path, file)
}
\arguments{
\item{path}{path to xlsx}

\item{file}{xml file corresponding to a specific worksheet}
}
\value{
an XML document
}
\description{
Read in the XML for a specific file within the xlsx, e.g. the file
corresponding to a specific worksheet.
}
\keyword{internal}

